Add computer setup

+++ pcforge-backend/src/main/java/com/kubacki/dawid/PCForge/models/components/ComputerCase.java
@@ -22,7 +22,7 @@ public class Case {
     private String motherboard;
 
     @Column(nullable=false)
-    private Double gpu_size;
+    private Integer gpu_size;
 
     @Column(nullable=false)
     private String power_supply;
